requirements.txt
requests==2.26.0
PyYAML==6.0